backend/requirements.txt
# Backend dependencies
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0

# File handling and utilities
python-multipart
python-dotenv
aiofiles
requests
PyJWT

# Google Gemini AI
google-generativeai>=0.8.0

# Google APIs (Calendar, Gmail, oAuth)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Supbase 
supabase>=2.0.0,<3.0.0
httpx>=0.27.0

# Email Notifications
jinja2
python-dateutil
pytz

# PDF Generation
# Note: Weasyprint requires system libs: brew install pango gdk-pixbuf libffi. See README.md for details.
markdown2>=2.4.0
weasyprint>=60.0,<70.0
